production: add quality integration domain and production gate
P3-B/C: QualityInspection 升级为可关联工序执行/报工的质量事实模型
- migration V20260819_003: execution_id/work_report_id + DECIMAL(18,4) 数量
- Quality Read Model: listByOrder/Execution/WorkReport + FQC 历史/PENDING/PASS 查询
- QualityActionService: judge/reinspect/createInspection + FQC 自动创建(最后 Execution 完成)
- Order gate: 仅最新 FQC PASS 可完成工单; finishedQuantity = FQC.passQuantity
- FQC FAIL 恢复 Execution 可继续生产; WorkReport cancel 与质检联动
- 已判定 PASS/FAIL 不可变; 复检新建记录不覆盖历史

+++ jjx-server/src/main/java/com/jjx/production/domain/dto/QualityInspectionUpdateDTO.java
@@ -1,13 +1,17 @@
 package com.jjx.production.domain.dto;
 import lombok.Data;
+import java.math.BigDecimal;
 import java.util.List;
 @Data
 public class QualityInspectionUpdateDTO {
     private Long inspectionId;
     private String result;
-    private Integer totalQty;
-    private Integer passQty;
-    private Integer failQty;
+    /** P3-B：实际检验数量（DECIMAL(18,4)） */
+    private BigDecimal totalQty;
+    /** P3-B：质量认可合格数量（DECIMAL(18,4)） */
+    private BigDecimal passQty;
+    /** P3-B：质量判定不合格数量（DECIMAL(18,4)） */
+    private BigDecimal failQty;
     private String defectDesc;
     private List<InspectionItemDTO> items;
 }
